+++ auth/token/maker.go
@@ -1,0 +1,8 @@
+package token
+
+import "time"
+
+type Maker interface {
+	CreateTocken(username string, duration time.Duration) (string, *Payload, error)
+	VerifyToken(token string) (*Payload, error)
+}
