--- token/jwt_maker.go
@@ -1,65 +1,0 @@
-package token
-
-import (
-	"errors"
-	"fmt"
-	"time"
-
-	"github.com/golang-jwt/jwt"
-)
-
-const minSecretKeySize = 32
-
-// error message
-var (
-	ErrInvalidToken = errors.New("token is invalid")
-	ErrExpiredToken = errors.New("token has expired")
-)
-
-type JWTMaker struct {
-	secretKey string
-}
-
-func NewJWTMaker(secretKey string) (Maker, error) {
-	if len(secretKey) < minSecretKeySize {
-		return nil, fmt.Errorf("key is short. key must be at latest%d characters", minSecretKeySize)
-	}
-	return &JWTMaker{secretKey}, nil
-}
-
-func (jm *JWTMaker) CreateTocken(username string, duration time.Duration) (string, *Payload, error) {
-	payload, err := NewPayload(username, duration)
-	if err != nil {
-		return "", payload, err
-	}
-
-	jwtToken := jwt.NewWithClaims(jwt.SigningMethodHS256, payload)
-	token, err := jwtToken.SignedString([]byte(jm.secretKey))
-	return token, payload, err
-}
-
-func (jm *JWTMaker) VerifyToken(token string) (*Payload, error) {
-	keyFunc := func(token *jwt.Token) (interface{}, error) {
-		_, ok := token.Method.(*jwt.SigningMethodHMAC)
-		if !ok {
-			return nil, ErrInvalidToken
-		}
-		return []byte(jm.secretKey), nil
-	}
-
-	jwtToken, err := jwt.ParseWithClaims(token, &Payload{}, keyFunc)
-	if err != nil {
-		verr, ok := err.(*jwt.ValidationError)
-		if ok && errors.Is(verr.Inner, ErrExpiredToken) {
-			return nil, ErrExpiredToken
-		}
-		return nil, ErrInvalidToken
-	}
-
-	payload, ok := jwtToken.Claims.(*Payload)
-	if !ok {
-		return nil, ErrInvalidToken
-	}
-
-	return payload, nil
-}
